Big Refactoring: deleted Service layout with implementation via JdbcTemplate , left Service layout with implementation via Hibernate. Renamed classes, moved to new directory
