+++ myBrikProject/src/main/java/by/mybrik/controllers/ProductTypeController.java
@@ -3,7 +3,7 @@ package by.mybrik.controllers;
 import by.mybrik.controllers.requests.productTypeRequests.ProductTypeCreate;
 import by.mybrik.controllers.requests.productTypeRequests.ProductTypeUpdate;
 import by.mybrik.domain.entities.ProductType;
-import by.mybrik.service.newImplementation.ProductTypeSer;
+import by.mybrik.service.ProductTypeService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.DeleteMapping;
@@ -24,7 +24,7 @@ import java.util.List;
 @RequiredArgsConstructor
 public class ProductTypeController {
 
-  public final ProductTypeSer typeService;
+  public final ProductTypeService typeService;
 
   // http://localhost:8080/new/rest/producttype
   @GetMapping
