+++ myBrikProject/src/main/java/by/mybrik/service/IndividualOrderService.java
@@ -1,22 +1,21 @@
 package by.mybrik.service;
 
-import by.mybrik.domain.IndividualOrder;
+import by.mybrik.domain.entities.IndividualOrder;
 
 import java.util.List;
 import java.util.Optional;
 
 public interface IndividualOrderService {
 
-    IndividualOrder save(IndividualOrder order);
+  IndividualOrder save(IndividualOrder order);
 
-    List<IndividualOrder> findAll();
+  List<IndividualOrder> findAll();
 
-    IndividualOrder findById(Long id);
+  IndividualOrder findById(Long id);
 
-    Optional<IndividualOrder> findOne(Long id);
+  Optional<IndividualOrder> findOne(Long id);
 
-    IndividualOrder update(IndividualOrder order);
+  IndividualOrder update(IndividualOrder order);
 
-    Long delete(IndividualOrder order);
-
+  Long delete(IndividualOrder order);
 }
