+++ myBrikProject/src/main/java/by/mybrik/service/PriceForIndividualOrderService.java
@@ -1,24 +1,22 @@
 package by.mybrik.service;
 
-
-import by.mybrik.domain.PriceForIndividualOrder;
+import by.mybrik.domain.entities.PriceForIndividualOrder;
 
 import java.util.List;
 import java.util.Optional;
 
 public interface PriceForIndividualOrderService {
 
-    PriceForIndividualOrder save(PriceForIndividualOrder price);
+  PriceForIndividualOrder save(PriceForIndividualOrder price);
 
-    List<PriceForIndividualOrder> findAll();
+  List<PriceForIndividualOrder> findAll();
 
-    PriceForIndividualOrder findById(Long id);
+  PriceForIndividualOrder findById(Long id);
 
-    Optional<PriceForIndividualOrder> findOne(Long id);
+  Optional<PriceForIndividualOrder> findOne(Long id);
 
-    PriceForIndividualOrder update(PriceForIndividualOrder price);
+  PriceForIndividualOrder update(PriceForIndividualOrder price);
 
-    Long delete(PriceForIndividualOrder price);
-
+  Long delete(PriceForIndividualOrder price);
 
 }
