+++ myBrikProject/src/main/java/by/mybrik/service/impl/GoodsServiceImpl.java
@@ -1,7 +1,7 @@
 package by.mybrik.service.impl;
 
-import by.mybrik.domain.Goods;
-import by.mybrik.repository.GoodsRepository;
+import by.mybrik.domain.entities.Goods;
+import by.mybrik.repository.newImplementation.GoodsRep;
 import by.mybrik.service.GoodsService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -13,7 +13,7 @@ import java.util.Optional;
 @RequiredArgsConstructor
 public class GoodsServiceImpl implements GoodsService {
 
-    private final GoodsRepository goodsRepository;
+    private final GoodsRep goodsRepository;
 
     @Override
     public Goods save(Goods product) {
@@ -44,4 +44,5 @@ public class GoodsServiceImpl implements GoodsService {
     public Long delete(Goods product) {
         return goodsRepository.delete(product);
     }
+
 }
