+++ myBrikProject/src/main/java/by/mybrik/service/impl/PriceForIndividualOrderServiceImpl.java
@@ -1,7 +1,7 @@
 package by.mybrik.service.impl;
 
-import by.mybrik.domain.PriceForIndividualOrder;
-import by.mybrik.repository.PriceForIndividualOrderRepository;
+import by.mybrik.domain.entities.PriceForIndividualOrder;
+import by.mybrik.repository.newImplementation.PriceForIndividualOrderRep;
 import by.mybrik.service.PriceForIndividualOrderService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -13,36 +13,35 @@ import java.util.Optional;
 @RequiredArgsConstructor
 public class PriceForIndividualOrderServiceImpl implements PriceForIndividualOrderService {
 
-    private final PriceForIndividualOrderRepository priceForIndividualOrderRepository;
+  private final PriceForIndividualOrderRep priceForIndividualOrderRepository;
 
+  @Override
+  public PriceForIndividualOrder save(PriceForIndividualOrder price) {
+    return priceForIndividualOrderRepository.save(price);
+  }
 
-    @Override
-    public PriceForIndividualOrder save(PriceForIndividualOrder price) {
-        return priceForIndividualOrderRepository.save(price);
-    }
+  @Override
+  public List<PriceForIndividualOrder> findAll() {
+    return priceForIndividualOrderRepository.findAll();
+  }
 
-    @Override
-    public List<PriceForIndividualOrder> findAll() {
-        return priceForIndividualOrderRepository.findAll();
-    }
+  @Override
+  public PriceForIndividualOrder findById(Long id) {
+    return priceForIndividualOrderRepository.findById(id);
+  }
 
-    @Override
-    public PriceForIndividualOrder findById(Long id) {
-        return priceForIndividualOrderRepository.findById(id);
-    }
+  @Override
+  public Optional<PriceForIndividualOrder> findOne(Long id) {
+    return priceForIndividualOrderRepository.findOne(id);
+  }
 
-    @Override
-    public Optional<PriceForIndividualOrder> findOne(Long id) {
-        return priceForIndividualOrderRepository.findOne(id);
-    }
+  @Override
+  public PriceForIndividualOrder update(PriceForIndividualOrder price) {
+    return priceForIndividualOrderRepository.update(price);
+  }
 
-    @Override
-    public PriceForIndividualOrder update(PriceForIndividualOrder price) {
-        return priceForIndividualOrderRepository.update(price);
-    }
-
-    @Override
-    public Long delete(PriceForIndividualOrder price) {
-        return priceForIndividualOrderRepository.delete(price);
-    }
+  @Override
+  public Long delete(PriceForIndividualOrder price) {
+    return priceForIndividualOrderRepository.delete(price);
+  }
 }
