--- myBrikProject/src/main/java/by/mybrik/service/newImplementation/impl/UsersSerImpl.java
@@ -1,47 +1,0 @@
-package by.mybrik.service.newImplementation.impl;
-
-import by.mybrik.domain.entities.Users;
-import by.mybrik.repository.newImplementation.UsersRep;
-import by.mybrik.service.newImplementation.UsersSer;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
-
-import java.util.List;
-import java.util.Optional;
-
-@Service
-@RequiredArgsConstructor
-public class UsersSerImpl implements UsersSer {
-
-    private final UsersRep usersRepository;
-
-    @Override
-    public Users save(Users user) {
-        return usersRepository.save(user);
-    }
-
-    @Override
-    public List<Users> findAll() {
-        return usersRepository.findAll();
-    }
-
-    @Override
-    public Users findById(Long id) {
-        return usersRepository.findById(id);
-    }
-
-    @Override
-    public Optional<Users> findOne(Long id) {
-        return usersRepository.findOne(id);
-    }
-
-    @Override
-    public Users update(Users user) {
-        return usersRepository.update(user);
-    }
-
-    @Override
-    public Long delete(Users user) {
-        return usersRepository.delete(user);
-    }
-}
